Helper\LocationHelper.php moved to Converter\LocationConverter.php
